* Adding support for Enhancement #6 - Add Support for Seeking to Specific Point in Timeline  - Click on ticker tape area moves all axes to position * Fixed bug with homing left of center

+++ Graffik/gui/film/MotionArea/motionpathpainter.h
@@ -86,7 +86,7 @@ public:
     float getFilmTime(int p_x);
     float getSpeed(unsigned long p_x);
     float getPosition(unsigned long p_x);
-    int getX(int p_time);
+    int getX(unsigned long p_time);
 
 signals:
